Add C FFI functions for table_row_get_ordinal_char_n and table_row_get_ordinal_date
