+++ packages/datajoint-core-ffi-c/src/results/mod.rs
@@ -1,2 +1,3 @@
 mod table_row_vector;
+mod table_row;
 pub use table_row_vector::TableRowVector;
